Fix up some things per review feedback
