+++ src/nunavut/lang/__init__.py
@@ -19,8 +19,6 @@ from ._language import LanguageClassLoader as LanguageClassLoader
 
 logger = logging.getLogger(__name__)
 
-NUNAVUT_LANG_CPP = "nunavut.lang.cpp"
-
 
 class UnsupportedLanguageError(ValueError):
     """
@@ -223,11 +221,10 @@ class LanguageContextBuilder:
         return self
 
     def load_default_config(self, language_standard: str) -> None:
-        self._ln_loader.config  # Accessing this property causes the defaults to load
-        defaults_key = f"{language_standard}_options"
-        if defaults_key in self._ln_loader.config.sections()[NUNAVUT_LANG_CPP]:
-            defaults_data = self._ln_loader.config.get_config_value_as_dict(NUNAVUT_LANG_CPP, defaults_key)
-            self._ln_loader.config.update_section(NUNAVUT_LANG_CPP, {"options": defaults_data})
+        self._ln_loader.config.apply_defaults(language_standard)
+
+    def validate_langauge_options(self) -> None:
+        self._ln_loader.config.validate_language_options()
 
     def set_additional_config_files(
         self, additional_config_files: typing.List[pathlib.Path]
