+++ src/nunavut/lang/cpp/templates/_composite_type.j2
@@ -33,52 +33,6 @@
 {% endif -%}
 struct {{composite_type|short_reference_name}} final
 {
-{%- if options.ctor_convention != ConstructorConvention.Default.value %}
-    using allocator_type = {{ options.allocator_type }}<decltype(nullptr)>;
-
-    // Default constructor
-    {{composite_type|short_reference_name}}(const allocator_type& allocator
-    {%- if options.allocator_is_default_constructible %} = allocator_type(){% endif %})
-    {%- if composite_type.fields_except_padding %} :{% endif %}
-    {%- for field in composite_type.fields_except_padding %}
-        {{ field | id }}{{ field | value_initializer(SpecialMethod.DefaultConstructorWithOptionalAllocator) }}{%if not loop.last %},{%endif %}
-    {%- endfor %}
-    { }
-
-    // Destructor
-    ~{{composite_type|short_reference_name}}() = default;
-
-    // Copy constructor
-    {{composite_type|short_reference_name}}(const {{composite_type|short_reference_name}}&) = default;
-
-    // Copy constructor with allocator
-    {{composite_type|short_reference_name}}(const {{composite_type|short_reference_name}}& rhs, const allocator_type& allocator)
-    {%- if composite_type.fields_except_padding %} :{% endif %}
-    {%- for field in composite_type.fields_except_padding %}
-        {{ field | id }}{{ field | value_initializer(SpecialMethod.CopyConstructorWithAllocator) }}{%if not loop.last %},{%endif %}
-    {%- endfor %}
-    {
-    }
-
-    // Move constructor
-    {{composite_type|short_reference_name}}({{composite_type|short_reference_name}}&&) = default;
-
-    // Move constructor with allocator
-    {{composite_type|short_reference_name}}({{composite_type|short_reference_name}}&& rhs, const allocator_type& allocator)
-    {%- if composite_type.fields_except_padding %} :{% endif %}
-    {%- for field in composite_type.fields_except_padding %}
-        {{ field | id }}{{ field | value_initializer(SpecialMethod.MoveConstructorWithAllocator) }}{%if not loop.last %},{%endif %}
-    {%- endfor %}
-    {
-    }
-
-    // Copy assignment
-    {{composite_type|short_reference_name}}& operator=(const {{composite_type|short_reference_name}}&) = default;
-
-    // Move assignment
-    {{composite_type|short_reference_name}}& operator=({{composite_type|short_reference_name}}&&) = default;
-{%- endif %}
-
     struct _traits_  // The name is surrounded with underscores to avoid collisions with DSDL attributes.
     {
         _traits_() = delete;
@@ -114,6 +68,9 @@ struct {{composite_type|short_reference_name}} final
         static_assert({# -#}
             ExtentBytes < (std::numeric_limits<{{ typename_unsigned_bit_length }}>::max() / 8U), {# -#}
             "This message is too large to be handled by the selected types");
+{% if options.ctor_convention != ConstructorConvention.Default.value %}
+        using allocator_type = {{ options.allocator_type }}<decltype(nullptr)>;
+{% endif %}
 {%- for field in composite_type.fields_except_padding %}
     {%- if loop.first %}
         struct TypeOf
@@ -126,6 +83,49 @@ struct {{composite_type|short_reference_name}} final
     {%- endif %}
 {%- endfor %}
     };
+{% if options.ctor_convention != ConstructorConvention.Default.value %}
+    // Default constructor
+    {{composite_type|short_reference_name}}(const _traits_::allocator_type& allocator
+    {%- if options.allocator_is_default_constructible %} = _traits_::allocator_type(){% endif %})
+    {%- if composite_type.fields_except_padding %} :{% endif %}
+    {%- for field in composite_type.fields_except_padding %}
+        {{ field | id }}{{ field | value_initializer(SpecialMethod.DefaultConstructorWithOptionalAllocator) }}{%if not loop.last %},{%endif %}
+    {%- endfor %}
+    { }
+
+    // Destructor
+    ~{{composite_type|short_reference_name}}() = default;
+
+    // Copy constructor
+    {{composite_type|short_reference_name}}(const {{composite_type|short_reference_name}}&) = default;
+
+    // Copy constructor with allocator
+    {{composite_type|short_reference_name}}(const {{composite_type|short_reference_name}}& rhs, const _traits_::allocator_type& allocator)
+    {%- if composite_type.fields_except_padding %} :{% endif %}
+    {%- for field in composite_type.fields_except_padding %}
+        {{ field | id }}{{ field | value_initializer(SpecialMethod.CopyConstructorWithAllocator) }}{%if not loop.last %},{%endif %}
+    {%- endfor %}
+    {
+    }
+
+    // Move constructor
+    {{composite_type|short_reference_name}}({{composite_type|short_reference_name}}&&) = default;
+
+    // Move constructor with allocator
+    {{composite_type|short_reference_name}}({{composite_type|short_reference_name}}&& rhs, const _traits_::allocator_type& allocator)
+    {%- if composite_type.fields_except_padding %} :{% endif %}
+    {%- for field in composite_type.fields_except_padding %}
+        {{ field | id }}{{ field | value_initializer(SpecialMethod.MoveConstructorWithAllocator) }}{%if not loop.last %},{%endif %}
+    {%- endfor %}
+    {
+    }
+
+    // Copy assignment
+    {{composite_type|short_reference_name}}& operator=(const {{composite_type|short_reference_name}}&) = default;
+
+    // Move assignment
+    {{composite_type|short_reference_name}}& operator=({{composite_type|short_reference_name}}&&) = default;
+{%- endif %}
 
 {%- for constant in composite_type.constants %}
     {% if loop.first %}
